проект разбит на модули
